scales text change learning mode

+++ src/com/ewareza/shapegame/app/utils/DimensionScaler.java
@@ -6,16 +6,16 @@ import com.ewareza.shapegame.resources.ScaledDimenRes;
 
 public class DimensionScaler {
     public static ViewGroup.LayoutParams getScaledParams(ViewGroup.LayoutParams params) {
-        params.width = ScaledDimenRes.getScaledDimenXForValue(params.width);
-        params.height = ScaledDimenRes.getScaledDimenYForValue(params.height);
+        params.width = (int) ScaledDimenRes.getScaledDimenXForValue(params.width);
+        params.height = (int) ScaledDimenRes.getScaledDimenYForValue(params.height);
 
         if(params instanceof LinearLayout.LayoutParams) {
             LinearLayout.LayoutParams linearLayoutParams = (LinearLayout.LayoutParams) params;
 
-            linearLayoutParams.leftMargin = ScaledDimenRes.getScaledDimenXForValue(linearLayoutParams.leftMargin);
-            linearLayoutParams.topMargin = ScaledDimenRes.getScaledDimenYForValue(linearLayoutParams.topMargin);
-            linearLayoutParams.rightMargin = ScaledDimenRes.getScaledDimenXForValue(linearLayoutParams.rightMargin);
-            linearLayoutParams.bottomMargin = ScaledDimenRes.getScaledDimenYForValue(linearLayoutParams.bottomMargin);
+            linearLayoutParams.leftMargin = (int) ScaledDimenRes.getScaledDimenXForValue(linearLayoutParams.leftMargin);
+            linearLayoutParams.topMargin = (int) ScaledDimenRes.getScaledDimenYForValue(linearLayoutParams.topMargin);
+            linearLayoutParams.rightMargin = (int) ScaledDimenRes.getScaledDimenXForValue(linearLayoutParams.rightMargin);
+            linearLayoutParams.bottomMargin = (int) ScaledDimenRes.getScaledDimenYForValue(linearLayoutParams.bottomMargin);
 
             return linearLayoutParams;
         }
